setup utils and constants folders for better organization of static data and animations
