+++ src/constants/footerData.ts
@@ -1,0 +1,27 @@
+export const footerData = [
+  {
+    title: "Company Info",
+    links: [
+      { name: "About Us", href: "#" },
+      { name: "Contact", href: "#" },
+      { name: "Careers", href: "#" },
+      { name: "Awards & Recognition", href: "#" },
+    ],
+  },
+  {
+    title: "Legal",
+    links: [
+      { name: "Privacy Policy", href: "#" },
+      { name: "Terms & Conditions", href: "#" },
+    ],
+  },
+  {
+    title: "Resources",
+    links: [
+      { name: "FAQs", href: "#" },
+      { name: "How it Works", href: "#" },
+      { name: "Partnerships", href: "#" },
+      { name: "Support Center", href: "#" },
+    ],
+  },
+];
