+++ src/constants/howItWorksData.ts
@@ -1,0 +1,17 @@
+export const steps = [
+  {
+    title: "Create an Account",
+    description:
+      "Sign up in minutes to get full access to our library and start browsing.",
+  },
+  {
+    title: "Explore & Select",
+    description:
+      "Browse a diverse selection of books and easily select your preferred titles with a single click.",
+  },
+  {
+    title: "Read & Return",
+    description:
+      "Start reading instantly on any device, then return the book with one click when you're done.",
+  },
+];
