+++ src/pages/HomePage/component/Footer.tsx
@@ -1,37 +1,8 @@
 import elibraryLogo from "../../../assets/elibrary_logo.png";
 import { FaFacebook, FaLinkedin, FaInstagram } from "react-icons/fa";
+import { footerData } from "../../../constants/footerData";
 
 const Footer = () => {
-  const footerData = [
-    {
-      title: "Company Info",
-      links: [
-        { name: "About Us", href: "#" },
-        { name: "Contact", href: "#" },
-        { name: "Careers", href: "#" },
-        { name: "Awards & Recognition", href: "#" },
-      ],
-    },
-
-    {
-      title: "Legal",
-      links: [
-        { name: "Privacy Policy", href: "#" },
-        { name: "Terms & Conditions", href: "#" },
-      ],
-    },
-
-    {
-      title: "Resources",
-      links: [
-        { name: "FAQs", href: "#" },
-        { name: "How it Works", href: "#" },
-        { name: "Partnerships", href: "#" },
-        { name: "Support Center", href: "#" },
-      ],
-    },
-  ];
-
   return (
     <footer className="py-4 bg-wrapper ">
       <div className="container mx-auto">
